address

+++ handle_p.c
@@ -1,9 +1,9 @@
 #include "main.h"
+#include <stdint.h>
 int handle_p(void *p)
 {
 	int m = 0;
-	unsigned int a = (unsigned int)p;
-
+	uintptr_t a = (uintptr_t)p;
 	m += handle(a, 16, 2);
 	return (m);
 }
